refactor : 리뷰 반영
by 용주님
- 사용하지 않는 LottoGame 삭제
- LottoMoney에서 LottoCount 클래스 분리
- LottoResult에 lottoMoney 멤버 변수 사용 대신 profit()에 lottoMoney 인자로 주입하는 것으로 대체
- 숫자에 _ 사용
- enum.values()로 순회
- utils 패키지 내 클래스명, 메서드명 수정
- LottoResultTest.java를 위한 생성자 추가
- 당첨번호 중복 검증 메서드 추가

by 유기님
- 패키지 컨벤션 소문자
- 생성자에서 비즈니스 로직 있는 경우 정적팩토리 메서드로 분리
- LottoTicket에서 StringToListInteger 변환로직 별도 클래스로 분리
- \n 대신 System.lineSeparator() 사용
- isInstanceOf 대신 isExactlyInstanceOf 사용
