--- src/main/java/Lotto/Lotto/LottoGame.java
@@ -1,5 +1,0 @@
-package Lotto.Lotto;
-
-public class LottoGame {
-
-}
