+++ src/main/java/lotto/lotto/LottoCount.java
@@ -1,0 +1,13 @@
+package lotto.lotto;
+
+public class LottoCount {
+    private final Integer count;
+
+    public LottoCount(Integer count) {
+        this.count = count;
+    }
+
+    public Integer getCount() {
+        return count;
+    }
+}
