+++ src/main/java/lotto/lotto/LottoTickets.java
@@ -1,0 +1,20 @@
+package lotto.lotto;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+public class LottoTickets {
+    private final List<LottoTicket> lottoTickets;
+
+    LottoTickets(LottoCount lottoCount) {
+        lottoTickets = IntStream.range(0, lottoCount.getCount())
+                .mapToObj(num -> new LottoTicket())
+                .collect(Collectors.toList());
+    }
+
+    public List<LottoTicket> getLottoTickets() {
+        return new ArrayList<>(lottoTickets);
+    }
+}
