+++ src/main/java/lotto/utils/RandomLottoNumbersGenerator.java
@@ -1,0 +1,16 @@
+package lotto.utils;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+public class RandomLottoNumbersGenerator {
+    private static final List<Integer> LOTTO_NUMBERS = IntStream.range(1, 46).boxed().collect(Collectors.toList());
+
+    public static List<Integer> generate() {
+        Collections.shuffle(LOTTO_NUMBERS);
+        return new ArrayList<>(LOTTO_NUMBERS.subList(0, 6));
+    }
+}
