+++ src/main/java/stringcalculator/calculator/Calculator.java
@@ -1,0 +1,17 @@
+package stringcalculator.calculator;
+
+import java.util.function.BinaryOperator;
+
+public class Calculator {
+    public static BinaryOperator<Float> add =
+            (number1, number2) -> number1 + number2;
+
+    public static BinaryOperator<Float> subtract =
+            (number1, number2) -> number1 - number2;
+
+    public static BinaryOperator<Float> multiply =
+            (number1, number2) -> number1 * number2;
+
+    public static BinaryOperator<Float> divide =
+            (number1, number2) -> number1 / number2;
+}
