--- src/test/java/Lotto/Lotto/LottoMoneyTest.java
@@ -1,16 +1,0 @@
-package Lotto.Lotto;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class LottoMoneyTest {
-    @ParameterizedTest
-    @CsvSource(value = {"10000,10", "14500,14", "900,0"}, delimiter = ',')
-    @DisplayName("돈으로 몇장의 티켓을 사는지 확인")
-    public void buyCountTest(Integer money, Integer ticketCount) {
-        assertThat(new LottoMoney(money).getLottoCount()).isEqualTo(ticketCount);
-    }
-}
